Moved constant to root.
